refactor: move post feed reads to post/api/post
